requirements.txt
Django==1.10.1
django-extensions==1.7.4
django-storages==1.5.1
django-taggit==0.21.2
Pillow==3.3.1
six==1.10.0
